HPCC-10992 Move getHomeDir to jmisc and update eclcc.cpp to use it
