+++ system/jlib/jmisc.cpp
@@ -27,8 +27,10 @@
 #include "jexcept.hpp"
 #include "jstring.hpp"
 #include "jdebug.hpp"
+
 #ifndef _WIN32
 #include <sys/wait.h>
+#include <pwd.h>
 #endif
 
 #ifdef LOGCLOCK
@@ -649,9 +651,6 @@ void close_program(HANDLE handle)
 
 #endif
 
-
-
-
 #ifndef _WIN32
 bool CopyFile(const char *file, const char *newfile, bool fail)
 {
@@ -936,4 +935,24 @@ StringBuffer & hexdump2string(byte const * in, size32_t inSize, StringBuffer & o
         out.appendf("x%u", seq);
     out.append(" ]");
     return out;
+}
+
+jlib_decl bool getHomeDir(StringBuffer & homepath)
+{
+#ifdef _WIN32
+    const char *home = getenv("APPDATA");
+    // Not the 'official' way - which changes with every windows version
+    // but should work well enough for us (and avoids sorting out windows include mess)
+#else
+    const char *home = getenv("HOME");
+    if (!home)
+    {
+        struct passwd *pw = getpwuid(getuid());
+        home = pw->pw_dir;
+    }
+#endif
+    if (!home)
+        return false;
+    homepath.append(home);
+    return true;
 }
