+++ system/jlib/jmisc.hpp
@@ -87,6 +87,7 @@ jlib_decl char* readarg(char*& curptr);
 jlib_decl bool invoke_program(const char *command_line, DWORD &runcode, bool wait=true, const char *outfile=NULL, HANDLE *rethandle=NULL, bool throwException = false, bool newProcessGroup = false);
 jlib_decl bool wait_program(HANDLE handle,DWORD &runcode,bool block=true);
 jlib_decl bool interrupt_program(HANDLE handle, bool killChildren, int signum=0); // no signum means use default
+jlib_decl bool getHomeDir(StringBuffer & homepath);
 
 #ifndef _WIN32
 jlib_decl bool CopyFile(const char *file, const char *newfile, bool fail);
